Changed all hashes to `size_t` and added constant error value; removed useless root name.

+++ src/filesystem_core.h
@@ -68,7 +68,7 @@ void fs__exit(void);
  * @pre   all the checks before the creation have already been made.
  * @post  the new file is now the head of the list of children starting at parent->content.l_child; if the hash table is expanded during the creation, *new_hash now contains the updated hash.
  */
-fs_file_t* fs__new(int* new_hash, char* new_name, bool is_dir, fs_file_t* parent);
+fs_file_t* fs__new(size_t* new_hash, char* new_name, bool is_dir, fs_file_t* parent);
 
 /**
  * Browse the filesystem following the path and return a pointer to the table cell identified by the path, creating a new file in such cell if requested.
